pyproject.toml
[project]
name = "kasir-api-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "psycopg[binary,pool]>=3.3.2",
    "python-dotenv>=1.2.1",
]
